+++ src/main/kotlin/com/github/bgomar/consolelogger/ConsoleLoggerRemove.kt
@@ -18,7 +18,7 @@ class ConsoleLoggerRemove : AnAction("Remove ConsoleLogger's Logs") {
     val project = e.getData(CommonDataKeys.PROJECT)!!
     val editor = e.getRequiredData(CommonDataKeys.EDITOR)
 
-    val patternToReplace = ".*" + ConsoleLoggerSettings.instance.pattern.run {
+    val removeLog: String = ".*" + ConsoleLoggerSettings.instance.defaultPattern.run {
       replace("\\", "\\\\")
         .replace("(", "\\(")
         .replace(")", "\\)")
@@ -39,8 +39,8 @@ class ConsoleLoggerRemove : AnAction("Remove ConsoleLogger's Logs") {
         .replace("$", "\\$")
     } + "\n"
 
-    val findModel = FindModel().apply {
-      stringToFind = patternToReplace
+    val findLog = FindModel().apply {
+      stringToFind = removeLog
       stringToReplace = ""
       isPromptOnReplace = false
       isRegularExpressions = true
@@ -48,10 +48,170 @@ class ConsoleLoggerRemove : AnAction("Remove ConsoleLogger's Logs") {
       isPromptOnReplace = false
     }
 
+    val removeInfo: String = ".*" + ConsoleLoggerSettings.instance.infoPattern.run {
+      replace("\\", "\\\\")
+        .replace("(", "\\(")
+        .replace(")", "\\)")
+        .replace("[", "\\[")
+        .replace("]", "\\]")
+        .replace("^", "\\^")
+        .replace("+", "\\+")
+        .replace("?", "\\?")
+        .replace("|", "\\|")
+        .replace(".", "\\.")
+        .replace("*", "\\*")
+        .replace("$$", ".*")
+        .replace("{FN}", ".*")
+        .replace("{FP}", ".*")
+        .replace("{LN}", "\\d*")
+        .replace("{", "\\{")
+        .replace("}", "\\}")
+        .replace("$", "\\$")
+    } + "\n"
+
+    val findInfo = FindModel().apply {
+      stringToFind = removeInfo
+      stringToReplace = ""
+      isPromptOnReplace = false
+      isRegularExpressions = true
+      isGlobal = true
+      isPromptOnReplace = false
+    }
+
+    val removeDebug: String = ".*" + ConsoleLoggerSettings.instance.debugPattern.run {
+      replace("\\", "\\\\")
+        .replace("(", "\\(")
+        .replace(")", "\\)")
+        .replace("[", "\\[")
+        .replace("]", "\\]")
+        .replace("^", "\\^")
+        .replace("+", "\\+")
+        .replace("?", "\\?")
+        .replace("|", "\\|")
+        .replace(".", "\\.")
+        .replace("*", "\\*")
+        .replace("$$", ".*")
+        .replace("{FN}", ".*")
+        .replace("{FP}", ".*")
+        .replace("{LN}", "\\d*")
+        .replace("{", "\\{")
+        .replace("}", "\\}")
+        .replace("$", "\\$")
+    } + "\n"
+
+    val findDebug = FindModel().apply {
+      stringToFind = removeDebug
+      stringToReplace = ""
+      isPromptOnReplace = false
+      isRegularExpressions = true
+      isGlobal = true
+      isPromptOnReplace = false
+    }
+
+    val removeWarn: String = ".*" + ConsoleLoggerSettings.instance.warnPattern.run {
+      replace("\\", "\\\\")
+        .replace("(", "\\(")
+        .replace(")", "\\)")
+        .replace("[", "\\[")
+        .replace("]", "\\]")
+        .replace("^", "\\^")
+        .replace("+", "\\+")
+        .replace("?", "\\?")
+        .replace("|", "\\|")
+        .replace(".", "\\.")
+        .replace("*", "\\*")
+        .replace("$$", ".*")
+        .replace("{FN}", ".*")
+        .replace("{FP}", ".*")
+        .replace("{LN}", "\\d*")
+        .replace("{", "\\{")
+        .replace("}", "\\}")
+        .replace("$", "\\$")
+    } + "\n"
+
+    val findWarn = FindModel().apply {
+      stringToFind = removeWarn
+      stringToReplace = ""
+      isPromptOnReplace = false
+      isRegularExpressions = true
+      isGlobal = true
+      isPromptOnReplace = false
+    }
+
+    val removeError: String = ".*" + ConsoleLoggerSettings.instance.errorPattern.run {
+      replace("\\", "\\\\")
+        .replace("(", "\\(")
+        .replace(")", "\\)")
+        .replace("[", "\\[")
+        .replace("]", "\\]")
+        .replace("^", "\\^")
+        .replace("+", "\\+")
+        .replace("?", "\\?")
+        .replace("|", "\\|")
+        .replace(".", "\\.")
+        .replace("*", "\\*")
+        .replace("$$", ".*")
+        .replace("{FN}", ".*")
+        .replace("{FP}", ".*")
+        .replace("{LN}", "\\d*")
+        .replace("{", "\\{")
+        .replace("}", "\\}")
+        .replace("$", "\\$")
+    } + "\n"
+    val findError = FindModel().apply {
+      stringToFind = removeError
+      stringToReplace = ""
+      isPromptOnReplace = false
+      isRegularExpressions = true
+      isGlobal = true
+      isPromptOnReplace = false
+    }
+    val removeTable: String = ".*" + ConsoleLoggerSettings.instance.tablePattern.run {
+      replace("\\", "\\\\")
+        .replace("(", "\\(")
+        .replace(")", "\\)")
+        .replace("[", "\\[")
+        .replace("]", "\\]")
+        .replace("^", "\\^")
+        .replace("+", "\\+")
+        .replace("?", "\\?")
+        .replace("|", "\\|")
+        .replace(".", "\\.")
+        .replace("*", "\\*")
+        .replace("$$", ".*")
+        .replace("{FN}", ".*")
+        .replace("{FP}", ".*")
+        .replace("{LN}", "\\d*")
+        .replace("{", "\\{")
+        .replace("}", "\\}")
+        .replace("$", "\\$")
+    } + "\n"
+    val findTable = FindModel().apply {
+      stringToFind = removeTable
+      stringToReplace = ""
+      isPromptOnReplace = false
+      isRegularExpressions = true
+      isGlobal = true
+      isPromptOnReplace = false
+    }
+
+
+
     when (dlg.scope) {
-      Scope.CURRENT_FILE -> FindUtil.replace(project, editor, 0, findModel)
+      Scope.CURRENT_FILE ->
+                FindUtil.replace(project, editor, 0, findLog) &&
+                FindUtil.replace(project, editor, 0, findInfo) &&
+                FindUtil.replace(project, editor, 0, findDebug) &&
+                FindUtil.replace(project, editor, 0, findWarn) &&
+                FindUtil.replace(project, editor, 0, findError) &&
+                FindUtil.replace(project, editor, 0, findTable)
       Scope.PROJECT -> {
-        ReplaceInProjectManager(project).replaceInPath(findModel)
+                ReplaceInProjectManager(project).replaceInPath(findLog);
+                ReplaceInProjectManager(project).replaceInPath(findInfo);
+                ReplaceInProjectManager(project).replaceInPath(findDebug);
+                ReplaceInProjectManager(project).replaceInPath(findWarn);
+                ReplaceInProjectManager(project).replaceInPath(findError);
+                ReplaceInProjectManager(project).replaceInPath(findTable);
       }
     }
   }