+++ src/main/kotlin/com/github/bgomar/consolelogger/UpdateNotify.kt
@@ -25,10 +25,10 @@ class UpdateNotify : StartupActivity {
     """
     <br/>
     In this version you'll find the ability<br><br>
-    👉 to insert the <b>filename</b> and <b>line number</b> into the log<br>
+    👉 to use  <b>console.</b> with  <b>log, info, debug, warn, error, table</b> press CTRL + ALT + (1 to 6)  <br>
     See ConsoleLogger Settings<br><br>
-    👉 to remove ConsoleLogger's logs from the <b>current file</b> or the <b>project</b><br>
-    See ConsoleLogger Remove command in <b>Code | Log It</b>
+    👉 to remove ConsoleLogger's logs from the <b>current file</b> or the <b>project</b>  press CTRL + ALT + 0<br>
+    See ConsoleLogger in <b>Code | ConsoleLogger</b>
     """
   }