+++ CHANGELOG.md
@@ -2,13 +2,30 @@
 
 ## [Unreleased]
 
+## [0.0.4]
+- shortcut for LOG insertion   (CTRL + ALT + 1)
+- shortcut for INFO insertion  (CTRL + ALT + 2)
+- shortcut for DEBUG insertion (CTRL + ALT + 3)
+- shortcut for WARN insertion  (CTRL + ALT + 4)
+- shortcut for ERROR insertion (CTRL + ALT + 5)
+- shortcut for TABLE insertion (CTRL + ALT + 6)
+- shortcut for deletion of all (CTRL + ALT + 0)
+-
+
+## [0.0.3]
+- Console.log ctrl+alt+1 console.warn ctrl+alt+4
+
 ## [0.0.2]
 - SOMETHING NEW
 
 ## [0.0.1]
 - BREAKING CHANGE: API class names are CamelCase. This replaces the  "lower  case version" naming.
 
-[Unreleased]: https://github.com/bg-omar/consolelogger/compare/v0.0.2...HEAD
+[Unreleased]: https://github.com/bg-omar/consolelogger/compare/v0.0.4...HEAD
+
+[0.0.4]: https://github.com/bg-omar/consolelogger/compare/v0.0.3...v0.0.4
+
+[0.0.3]: https://github.com/bg-omar/consolelogger/compare/v0.0.2...v0.0.3
 
 [0.0.2]: https://github.com/bg-omar/consolelogger/compare/v0.0.1...v0.0.2